ref: update profile feature
