+++ src/hooks/useProfile.ts
@@ -1,22 +1,29 @@
 import { PROFILE_ENDPOINT } from '@/const/api';
-import { Profile, profileService } from '@/services/profileService';
+import { profileService } from '@/services/profileService';
 import useProfileStore from '@/store/useProfileStore'
+import useSidenavbarStore from '@/store/useSidenavbarStore';
+import { Profile } from '@/types/profileTypes';
 import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
 export const useFetchProfile = () => {
-    const { fetchProfile } = useProfileStore();
+    const { fetchProfile, profile } = useProfileStore();
+    const { setTitleToTopBar } = useSidenavbarStore();
 
     const query = useQuery<Profile, Error>({
         queryKey: [PROFILE_ENDPOINT],
         queryFn: () => profileService.getProfile(),
+        enabled: !profile,
         staleTime: 5 * 60 * 1000,
     });
 
     // Necesario para no entrar en bucle
     useEffect(() => {
-        query.data && fetchProfile(query.data);
-    }, [query.data, fetchProfile])
+        if (query.data && !profile) {
+            fetchProfile(query.data)
+            setTitleToTopBar(query.data.store.name)
+        }
+    }, [query.data, fetchProfile, profile])
 
     return query;
 }
