+++ src/types/profileTypes.ts
@@ -1,0 +1,21 @@
+export interface Profile {
+    id: string;
+    name: string;
+    last_name: string;
+    birthdate: string;
+    email: string;
+    status: string;
+    createdAt: Date;
+    updatedAt: Date;
+    store: Store; // Un store, no un array
+}
+
+export interface Store {
+    id: string;
+    name: string;
+    tel: string;
+    address: string;
+    createdAt: Date;
+    updatedAt: Date;
+    status: string;
+}
